レポート一括公開でconfigIdを直接使用するように修正 (#731)
billIdからconfigIdを解決する際に.single()を使用しており、
1つの議案に複数のインタビュー設定がある場合にエラーになっていた。
URLパラメータのconfigIdを直接使用し、bill/config整合性チェックを追加。
